refactor: Header Components and hide header options
