+++ Frontend/justina.io/src/components/Show/Show.jsx
@@ -1,0 +1,3 @@
+const Show = ({ when, children, fallback }) => (when ? children : fallback);
+
+export default Show;
